+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # domino x.x.x (not yet released)
 * Fix "#id" optimization in querySelectorAll() when 0 or 2 matches for
   `id`. (#99)
+* Correct return value of CSSStyleDeclaration#getPropertyValue() when
+  style is not set. (#98)
 
 # domino 1.0.28 (27 Jan 2017)
 * Fix unescape mechanism in attribute values. (#95)